cli/pyproject.toml
[project]
name = "sdlc-factory"
version = "1.0.0"
description = "SDLC-Factory cli tool."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.70.0",
    "jsonschema>=4.26.0",
    "mcp>=1.27.0",
    "openinference-instrumentation>=0.1.46",
    "openinference-instrumentation-google-genai>=0.1.15",
    "opentelemetry-exporter-otlp>=1.41.0",
    "opentelemetry-sdk>=1.41.0",
    "pgvector>=0.4.2",
    "psycopg[binary]>=3.3.3",
    "tenacity>=9.1.4",
    "typer>=0.24.1",
]
[project.scripts]
sdlc-factory = "sdlc_factory.main:app"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "pytest-mock>=3.15.1",
]
